rename activity package to view
