+++ app/src/main/java/com/javahelp/frontend/view/ClientRegistrationActivity.java
@@ -1,4 +1,4 @@
-package com.javahelp.frontend.activity;
+package com.javahelp.frontend.view;
