+++ app/src/main/java/com/javahelp/frontend/view/ProviderRegistrationViewmodel.java
@@ -1,4 +1,4 @@
-package com.javahelp.frontend.activity;
+package com.javahelp.frontend.view;
 
 import android.app.Application;
 
@@ -12,8 +12,8 @@ import com.javahelp.frontend.domain.user.register.RegisterInteractor;
 import com.javahelp.frontend.domain.user.register.RegisterResult;
 import com.javahelp.frontend.gateway.LambdaRegisterDataAccess;
 import com.javahelp.model.token.Token;
-import com.javahelp.model.user.ClientUserInfo;
 import com.javahelp.model.user.IPasswordHasher;
+import com.javahelp.model.user.ProviderUserInfo;
 import com.javahelp.model.user.SHAPasswordHasher;
 import com.javahelp.model.user.User;
 
@@ -21,23 +21,23 @@ import java.util.Optional;
 import java.util.concurrent.ExecutorService;
 import java.util.concurrent.Executors;
 
-public class ClientRegistrationVm extends AndroidViewModel implements IRegisterOutput {
+public class ProviderRegistrationViewmodel extends AndroidViewModel implements IRegisterOutput {
 
     private MutableLiveData<String> username = new MutableLiveData<>("");
     private MutableLiveData<String> password1 = new MutableLiveData<>("");
     private MutableLiveData<String> password2 = new MutableLiveData<>("");
     private MutableLiveData<String> email = new MutableLiveData<>("");
-    private MutableLiveData<String> firstname = new MutableLiveData<>("");
-    private MutableLiveData<String> lastname = new MutableLiveData<>("");
+    private MutableLiveData<String> practicename = new MutableLiveData<>("");
     private MutableLiveData<String> address = new MutableLiveData<>("");
     private MutableLiveData<String> phone = new MutableLiveData<>("");
+    private MutableLiveData<Boolean> certified = new MutableLiveData<>(Boolean.FALSE);
     private ExecutorService executor = Executors.newSingleThreadExecutor();
     private MutableLiveData<Optional<RegisterResult>> registerResult = new MutableLiveData<>(Optional.empty());
-    private ClientUserInfo clientUserInfo = new ClientUserInfo("", "", "", "", "");
+    private ProviderUserInfo providerUserInfo = new ProviderUserInfo("", "", "", "");
     private RegisterInteractor registerInteractor;
 
-    public ClientRegistrationVm(@NonNull Application application) {
 
+    public ProviderRegistrationViewmodel(@NonNull Application application) {
         super(application);
         IRegisterDataAccess register = LambdaRegisterDataAccess.getInstance();
         IPasswordHasher hasher = SHAPasswordHasher.getInstance();
@@ -45,7 +45,7 @@ public class ClientRegistrationVm extends AndroidViewModel implements IRegisterO
     }
 
     /**
-     * Sets the username for this {@link ClientRegistrationVm}
+     * Sets the username for this {@link ProviderRegistrationViewmodel}
      *
      * @param username {@link String} username
      */
@@ -54,7 +54,7 @@ public class ClientRegistrationVm extends AndroidViewModel implements IRegisterO
     }
 
     /**
-     * Sets the password for this {@link ClientRegistrationVm}
+     * Sets the password for this {@link ProviderRegistrationViewmodel}
      *
      * @param password1 {@link String} password1 to use
      */
@@ -63,7 +63,7 @@ public class ClientRegistrationVm extends AndroidViewModel implements IRegisterO
     }
 
     /**
-     * Sets the confirmed_password for this {@link ClientRegistrationVm}
+     * Sets the confirmed_password for this {@link ProviderRegistrationViewmodel}
      *
      * @param password2 {@link String} password2 to use
      */
@@ -72,7 +72,7 @@ public class ClientRegistrationVm extends AndroidViewModel implements IRegisterO
     }
 
     /**
-     * Sets the email for this {@link ClientRegistrationVm}
+     * Sets the email for this {@link ProviderRegistrationViewmodel}
      *
      * @param email {@link String} email to use
      */
@@ -81,25 +81,16 @@ public class ClientRegistrationVm extends AndroidViewModel implements IRegisterO
     }
 
     /**
-     * Sets the firstname for this {@link ClientRegistrationVm}
+     * Sets the practicename for this {@link ProviderRegistrationViewmodel}
      *
-     * @param firstname {@link String} firstname to use
+     * @param practicename {@link String} firstname to use
      */
-    public void setFirstname(String firstname) {
-        this.firstname.setValue(firstname);
+    public void setPracticename(String practicename) {
+        this.practicename.setValue(practicename);
     }
 
     /**
-     * Sets the lastname for this {@link ClientRegistrationVm}
-     *
-     * @param lastname {@link String} firstname to use
-     */
-    public void setLastname(String lastname) {
-        this.firstname.setValue(lastname);
-    }
-
-    /**
-     * Sets the address for this {@link ClientRegistrationVm}
+     * Sets the address for this {@link ProviderRegistrationViewmodel}
      *
      * @param address {@link String} address to use
      */
@@ -108,12 +99,21 @@ public class ClientRegistrationVm extends AndroidViewModel implements IRegisterO
     }
 
     /**
-     * Sets the phone for this {@link ClientRegistrationVm}
+     * Sets the phone for this {@link ProviderRegistrationViewmodel}
      *
      * @param phone {@link String} phone to use
      */
     public void setPhone(String phone) {
         this.phone.setValue(phone);
+    }
+
+    /**
+     * Sets whether the current provider is certified or not
+     *
+     * @param c whether certified
+     */
+    public void setCertified(boolean c) {
+        certified.setValue(c);
     }
 
     /**
@@ -126,16 +126,16 @@ public class ClientRegistrationVm extends AndroidViewModel implements IRegisterO
     }
 
     /**
-     * Sets the clientUserInfo for this {@link ClientRegistrationVm}
+     * Sets the provideruserinfo for this {@link ProviderRegistrationViewmodel}
      *
-     * @param email, firstname, lastname, phone, address {@link String}
+     * @param email, practicename, phone, address {@link String}
      */
-    public void setClientUserInfo(String email, String address, String phone, String firstname, String lastname) {
-        this.clientUserInfo.setEmailAddress(email);
-        this.clientUserInfo.setFirstName(firstname);
-        this.clientUserInfo.setLastName(lastname);
-        this.clientUserInfo.setAddress(address);
-        this.clientUserInfo.setPhoneNumber(phone);
+    public void setProviderUserInfo(String email, String practicename, String phone, String address) {
+        this.providerUserInfo.setEmailAddress(email);
+        this.providerUserInfo.setPracticeName(practicename);
+        this.providerUserInfo.setCertified(false);
+        this.providerUserInfo.setAddress(address);
+        this.providerUserInfo.setPhoneNumber(phone);
     }
 
     /**
@@ -144,7 +144,6 @@ public class ClientRegistrationVm extends AndroidViewModel implements IRegisterO
     public MutableLiveData<Optional<RegisterResult>> getRegisterResult() {
         return registerResult;
     }
-
 
     /**
      * @param password1, password2 {@link String}
@@ -155,28 +154,37 @@ public class ClientRegistrationVm extends AndroidViewModel implements IRegisterO
 
     }
 
+
     /**
      * Tries to Register in
      */
     public void attemptRegister() {
         executor.execute(() -> {
             try {
+                // the delay is present in case login immediately returns an error.
+                // in the event that login immediately returns an error, if the user does not
+                // see the error message under the login button, they may not realize anything
+                // has happened. Instead, it is easier to add a small delay so that the user
+                // sees the program 'trying' to log in, sees the progress bar appear, sees it
+                // disappear, and understands that the login has failed.
                 Thread.sleep(250);
             } catch (InterruptedException ignored) {
 
             }
             if (passwordMatch(password1.getValue(), password2.getValue())) {
-                ClientUserInfo info = new ClientUserInfo(
+                ProviderUserInfo info = new ProviderUserInfo(
                         email.getValue(),
                         address.getValue(),
                         phone.getValue(),
-                        firstname.getValue(),
-                        lastname.getValue());
+                        practicename.getValue());
+                info.setCertified(certified.getValue());
                 User u = new User("", info, username.getValue());
                 registerInteractor.register(u, password1.getValue());
             }
+
         });
     }
+
 
     @Override
     public void success(User user, Token token) {
