+++ app/src/main/res/layout/activity_client_registration.xml
@@ -6,7 +6,7 @@
     <data>
         <variable
             name="cdata"
-            type="com.javahelp.frontend.activity.ClientRegistrationVm" />
+            type="com.javahelp.frontend.view.ClientRegistrationVm" />
 
     </data>
 
@@ -15,7 +15,7 @@
         android:layout_height="match_parent"
         android:background="@color/white"
         android:orientation="vertical"
-        tools:context=".frontend.activity.ProviderRegistrationActivity">
+        tools:context=".frontend.view.ProviderRegistrationActivity">
 
         <ScrollView
             android:layout_width="match_parent"
