+++ app/src/main/res/layout/activity_delete_verify.xml
@@ -6,7 +6,7 @@
     <data>
         <variable
             name="data"
-            type="com.javahelp.frontend.activity.DeleteViewModel" />
+            type="com.javahelp.frontend.view.DeleteViewModel" />
 
     </data>
 
@@ -16,7 +16,7 @@
         android:orientation="vertical"
         android:padding="10dp"
         android:background="@drawable/b2"
-        tools:context=".frontend.activity.DeleteActivity">
+        tools:context=".frontend.view.DeleteActivity">
 
 
         <TextView
