+++ app/src/main/res/layout/activity_fg.xml
@@ -6,7 +6,7 @@
     <data>
         <variable
             name="data"
-            type="com.javahelp.frontend.activity.FrontPageActivityVm" />
+            type="com.javahelp.frontend.view.FrontPageActivityVm" />
 
     </data>
 
@@ -14,7 +14,7 @@
         android:layout_width="match_parent"
         android:orientation="vertical"
         android:layout_height="match_parent"
-        tools:context=".frontend.activity.FrontPageActivity">
+        tools:context=".frontend.view.FrontPageActivity">
         <FrameLayout
             android:id="@+id/frame_layout"
             android:layout_width="0dp"
