+++ app/src/main/res/layout/activity_front_page2.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="vertical"
-    tools:context=".frontend.activity.FrontPage2Activity">
+    tools:context=".frontend.view.FrontPage2Activity">
 
     <TextView
         android:layout_width="match_parent"
